casting works, changing all

+++ src/convert.c
@@ -170,18 +170,18 @@ char* JpmcdsFormatDate(TDate date) /* (I) */
         }
         else if (mdy.month < 10 && mdy.day >= 10)
         {
-            snprintf(&format[ibuf][0], MAX_STR_LEN, "%04ld%02ld%02ld",
-                    mdy.year, mdy.month, mdy.day );
+            snprintf(&format[ibuf][0], MAX_STR_LEN, "%04d%02d%02d",
+                     (int)mdy.year, (int)mdy.month, (int)mdy.day );
         }
         else if (mdy.month >= 10 && mdy.day < 10)
         {
-            snprintf(&format[ibuf][0], MAX_STR_LEN, "%04ld%02ld%02ld",
-                    mdy.year, mdy.month, mdy.day );
+            snprintf(&format[ibuf][0], MAX_STR_LEN, "%04d%02d%02d",
+                     (int)mdy.year, (int)mdy.month, (int)mdy.day );
         }
         else   /* month && day >= 10 */
         {
-            snprintf(&format[ibuf][0], MAX_STR_LEN, "%04ld%02ld%02ld",
-                    mdy.year, mdy.month, mdy.day );
+            snprintf(&format[ibuf][0], MAX_STR_LEN, "%04d%02d%02d",
+                     (int)mdy.year, (int)mdy.month, (int)mdy.day );
         }
     }
     return &format[ibuf][0];
